+++ Sources/SmileID/Classes/Views/ImageCaptureConfirmationDialog.swift
@@ -27,8 +27,8 @@ struct ImageCaptureConfirmationDialog: View {
             Image(uiImage: image)
                 .resizable()
                 .aspectRatio(contentMode: .fit)
-                .frame(height: 256)
                 .cornerRadius(16)
+                .frame(height: 256)
                 .scaleEffect(scaleFactor, anchor: .center)
                 .clipped()
             VStack(spacing: 16) {